ORB Maintenance 1.7 - Renamed Core.h to Base.h
[project]
- Renamed OrbE/Core/Core.h to OrbE/Core/Base.h.

[engine]
- Renamed the uniform sampler2D to u_Textures.
- Was testing the option of branching instead of indexing the texture sampler, but decided to stay with indexing as it's honestly less code duplication, not sure about the impact of branching in the fragment shader.

+++ SandOrb/src/SandOrb2D.cpp
@@ -83,7 +83,7 @@ void SandOrb2D::OnAttach()
 	m_PacWorldSubTexture = ORB::SubTexture2D::CreateFromRECT_BL(m_PacmanTexture, { 0.0f, 32.0f, 1024.0f, 768.0f });
 
 	m_PacmanMapWidth = s_PacmanMapWidth;
-	m_PacmanMapHeight = strlen(s_PacmanMapTiles) / s_PacmanMapWidth;
+	m_PacmanMapHeight = (uint32_t)(strlen(s_PacmanMapTiles)) / s_PacmanMapWidth;
 
 	m_PacmanTextureMap['.'] = ORB::SubTexture2D::CreateFromRECT_BL(m_PacmanTexture, {32.0f, 0.0f, 32.0f, 32.0f});
 	m_PacmanTextureMap['o'] = ORB::SubTexture2D::CreateFromRECT_BL(m_PacmanTexture, {0.0f, 0.0f, 32.0f, 32.0f});
